thesis-docs/requirements-irb.txt
# For parsing IRB Word documents
python-docx>=1.1.0
